+++ definitions/luxon-ext.d.ts
@@ -1,0 +1,7 @@
+import * as luxon from "luxon";
+
+declare module "luxon" {
+    interface DateTime {
+        toRelative(params?: { date?: DateTime }): string;
+    }
+}
